+++ tests/apps/another_app/apps.py
@@ -5,3 +5,4 @@ class AnotherAppConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
     name = "apps.another_app"
     label = "another_app"
+    has_reactive_components = True